add project entity, relation many to many, many to one, one to many, timesheet, filtering with used budget

+++ src/main/java/com/example/demo/user/dto/UserDto.java
@@ -6,6 +6,7 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -29,6 +30,6 @@ public class UserDto {
 
     private final String email;
 
-    private final int salaryPerHour;
+    private final BigDecimal salaryPerHour;
 
 }
